+++ CHANGE-LOG.md
@@ -1,7 +1,17 @@
-## Version 0.9.1 
+## Version 0.9.4 (released 2020-05-16)
+- Add wanted bytes at time parameter for stream encryption/decryption (**FEATURE**)
+- Add async versions of stream processing (**FEATURE**)
+
+## Version 0.9.3 (released 2020-05-09)
+- Support stream encryption/decryption (**FEATURE**)
+
+## Version 0.9.2 (released 2020-01-12)
+- Generate documentation (**FEATURE**)
+
+## Version 0.9.1 (released 2018-11-18)
 - More overloads for encrypt/decrypt (no need to pass lengths of arrays anymore)
 - One method to handle strings (as UTF-8 byte array)
  
-## Version 0.9.0
+## Version 0.9.0 (released 2018-11-14)
 - Initial Nuget release
 - Just one encrypt and decrypt method